requirements.txt
black==24.4.2
pytest==8.3.2
ruff==0.5.7
pyyaml==6.0.2
